requirements.txt
asn1crypto==0.22.0
bcrypt==3.1.3
certifi==2017.7.27.1
cffi==1.10.0
chardet==3.0.4
click==6.7
clickclick==1.2.2
connexion==1.4.2
cryptography==2.0.3
cycler==0.10.0
Flask==0.12.2
future==0.16.0
idna==2.5
inflection==0.3.1
itsdangerous==0.24
Jinja2==2.9.6
jsonschema==2.6.0
MarkupSafe==1.0
matplotlib==2.0.2
pathlib==1.0.1
psutil==5.2.2
pyasn1==0.3.2
pycparser==2.18
PyNaCl==1.1.2
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2017.2
PyYAML==3.12
requests==2.18.3
scipy==0.19.1
seaborn==0.8
six==1.10.0
swagger-spec-validator==2.1.0
typing==3.6.2
urllib3==1.22
Werkzeug==0.12.2
